added room database for ref data storage

+++ app/src/main/java/com/humara/nagar/ui/signup/model/AppConfigResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class AppConfigResponse(
     @SerializedName("role") val role: String?,
     @SerializedName("ward") val ward: String,
-    @SerializedName("wardId") val wardId: Long
+    @SerializedName("wardId") val wardId: Int
 )
 
 enum class Role(val role: String) {
